// src/types/clinic.d.ts
	interface IDoctor {
		doctorId: string;
		name: string;
		degree: string;
		speciality: string;
		experience: number;
		profileUrl?: string;
		mobile?: string;
		active: boolean;
	}

	interface IClinicDoctor {
		_id: string;
		clinicId: string;
		active: boolean;
		loginAccess: string;
		doctor: IDoctor;
		createdAtIST?: string | null;
	}